Move conditionfuncs to commitinfo
